OTELのSDKの初期化をレイヤー内のCJSで行うように変更
